refactor: account package
