+++ backend/src/main/java/org/tarjetamish/card/repository/impl/JdbcCardRepository.java
@@ -6,7 +6,6 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 import org.tarjetamish.card.model.Card;
 import org.tarjetamish.card.repository.CardRepository;
-import org.tarjetamish.card.mapper.CardRowMapper;
 
 import java.util.List;
 import java.util.Optional;
